Updates for 2.0.3: Fix caching of manifest
Bug: React on Rails tests would load the manifest, then detect that we
need to compile, and then WebpackerLite would fail to reload the
manifest file. Either the old JS files would be used, or, if the test
build script deletes the compiled files, then the test would fail from
the named files being non-existent.

Fix: The FileLoader tracks the mtime of any file loaded, and a call to
load_instance will check that mtime if not in production.

Other changes:
* Tests improved. Expected and actual params were reversed.
* Support for an env value to use the pry debugger.